app/multi_modal/handlers.py
import base64
import io
from PIL import Image
from typing import Dict, Any, List
import logging

logger = logging.getLogger(__name__)

class MultiModalHandler:
    def __init__(self):
        self.supported_formats = ["jpeg", "jpg", "png", "gif", "bmp"]
    
    def process_image(self, image_data: str, image_type: str) -> Dict[str, Any]:
        """Process base64 encoded image"""
        try:
            # Decode base64 image
            image_bytes = base64.b64decode(image_data)
            image = Image.open(io.BytesIO(image_bytes))
            
            # Convert to RGB if necessary
            if image.mode != 'RGB':
                image = image.convert('RGB')
            
            # Get image metadata
            metadata = {
                "format": image.format,
                "size": image.size,
                "mode": image.mode,
                "type": image_type
            }
            
            return {
                "image": image,
                "metadata": metadata,
                "processed": True
            }
            
        except Exception as e:
            logger.error(f"Error processing image: {e}")
            return {
                "image": None,
                "metadata": {},
                "processed": False,
                "error": str(e)
            }
    
    def extract_text_from_image(self, image: Image.Image) -> str:
        """Extract text from image using OCR (placeholder)"""
        # In a real implementation, you would use OCR library like pytesseract
        # For now, return placeholder text
        return "OCR text extraction not implemented"
    
    def analyze_image_content(self, image: Image.Image) -> Dict[str, Any]:
        """Analyze image content (placeholder)"""
        # In a real implementation, you would use vision models
        return {
            "objects": [],
            "scene": "unknown",
            "text": self.extract_text_from_image(image),
            "analysis": "Image analysis not implemented"
        }
