requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
openai>=1.10.0,<2.0.0
langchain==0.1.0
langchain-openai==0.0.5
google-generativeai==0.3.2
groq==0.4.1
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
redis==5.0.1
asyncio==3.4.3
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==0.8.11